+++ front/tsconfig.json
@@ -22,7 +22,6 @@
       "@utils/*": ["common/utils/*"],
       "@userContainer/*": ["common/userContainer/*"],
       "@modal/*": ["common/modal/*"],
-      "@recaptcha/*": ["common/recaptcha/*"],
       "@footer/*": ["common/footer/*"],
       "@header/*": ["common/header/*"],
       "@sidebar/*": ["common/sidebar/*"],
